@@ -25,7 +25,7 @@ import { ItemType } from '../entity/item-type';
 import { Publisher } from '../entity/publisher';
 import { Team } from '../entity/Team';
 import { TaskOrchestra } from '../task/task-orchestra';
-import { ConfigLoader, ItemStorage, TYPES } from '../types';
+import { ConfigLoader, ItemStorage, TYPES_IDX } from '../TYPES_IDX';
 import { toUTCDate, trimDomain } from '../utils/normalize';
 import { captureException } from '../utils/sentry';
 import { BaseScraper } from './abstract/base-scraper';
@@ -77,9 +77,9 @@ export class DmhyScraper extends BaseScraper<number> {
     private static _host = 'https://share.dmhy.org';
     private _browser: Browser;
 
-    constructor(@inject(TYPES.ItemStorage) store: ItemStorage<number>,
+    constructor(@inject(TYPES_IDX.ItemStorage) store: ItemStorage<number>,
                 @inject(TaskOrchestra) taskOrchestra: TaskOrchestra,
-                @inject(TYPES.ConfigLoader) config: ConfigLoader) {
+                @inject(TYPES_IDX.ConfigLoader) config: ConfigLoader) {
         super(taskOrchestra, config, store);
     }
 
